# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pottery & Artisan eCommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  images: string[];
  stock: number;
  colors?: string[];
  /** @nullable */
  handmadeDetails?: string | null;
  isFeatured?: boolean;
  createdAt: string;
}

export interface ProductList {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  categoryId: number;
  images: string[];
  /** @minimum 0 */
  stock: number;
  colors?: string[];
  handmadeDetails?: string;
  isFeatured?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  categoryId?: number;
  images?: string[];
  /** @minimum 0 */
  stock?: number;
  colors?: string[];
  handmadeDetails?: string;
  isFeatured?: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  image?: string;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerCity: string;
  customerAddress: string;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  totalPrice: number;
  status: string;
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  customerCity: string;
  /** @minLength 1 */
  customerAddress: string;
  notes?: string;
  items: OrderItemInput[];
}

export interface OrderStatusUpdate {
  status: string;
}

export interface OrderList {
  items: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface CsvExport {
  csv: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  username: string;
  authenticated: boolean;
}

export interface StatsSummary {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  pendingOrders: number;
  recentOrders: Order[];
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalOrdered: number;
}

export interface OrderStatusCount {
  status: string;
  count: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
minPrice?: number;
maxPrice?: number;
inStock?: boolean;
page?: number;
limit?: number;
};

export type ListOrdersParams = {
status?: string;
page?: number;
limit?: number;
};