# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pottery & Artisan eCommerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: orders
    description: Order operations
  - name: admin
    description: Admin operations
  - name: stats
    description: Statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ─── Products ───────────────────────────────────────────────────────────────
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filtering
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: inStock
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 12 }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"

    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/bestseller products for homepage hero
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

    patch:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products, admin]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Categories ─────────────────────────────────────────────────────────────
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories, admin]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # ─── Orders ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders, admin]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Submit a WhatsApp order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders, admin]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

    patch:
      operationId: updateOrderStatus
      tags: [orders, admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/export/csv:
    get:
      operationId: exportOrdersCsv
      tags: [orders, admin]
      summary: Export all orders as CSV
      responses:
        "200":
          description: CSV file content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvExport"

  # ─── Admin Auth ──────────────────────────────────────────────────────────────
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session info
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized

  # ─── Stats (wow endpoints) ────────────────────────────────────────────────
  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats, admin]
      summary: Admin dashboard stats — total orders, revenue, products, recent activity
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/top-products:
    get:
      operationId: getTopProducts
      tags: [stats, admin]
      summary: Most ordered products
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"

  /stats/orders-by-status:
    get:
      operationId: getOrdersByStatus
      tags: [stats, admin]
      summary: Order count grouped by status
      responses:
        "200":
          description: Orders by status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderStatusCount"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }

    # ── Product ──────────────────────────────────────────────────────────────
    Product:
      type: object
      required: [id, name, description, price, categoryId, images, stock, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        categoryId: { type: integer }
        categoryName: { type: ["string", "null"] }
        images:
          type: array
          items: { type: string }
        stock: { type: integer }
        colors:
          type: array
          items: { type: string }
        handmadeDetails: { type: ["string", "null"] }
        isFeatured: { type: boolean }
        createdAt: { type: string }

    ProductList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    ProductInput:
      type: object
      required: [name, description, price, categoryId, images, stock]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        categoryId: { type: integer }
        images:
          type: array
          items: { type: string }
        stock: { type: integer, minimum: 0 }
        colors:
          type: array
          items: { type: string }
        handmadeDetails: { type: string }
        isFeatured: { type: boolean }

    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        categoryId: { type: integer }
        images:
          type: array
          items: { type: string }
        stock: { type: integer, minimum: 0 }
        colors:
          type: array
          items: { type: string }
        handmadeDetails: { type: string }
        isFeatured: { type: boolean }

    # ── Category ─────────────────────────────────────────────────────────────
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        image: { type: ["string", "null"] }

    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string }
        image: { type: string }

    # ── Order ────────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, customerName, customerPhone, customerCity, customerAddress, items, totalPrice, status, createdAt]
      properties:
        id: { type: integer }
        customerName: { type: string }
        customerPhone: { type: string }
        customerCity: { type: string }
        customerAddress: { type: string }
        notes: { type: ["string", "null"] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice: { type: number }
        status: { type: string }
        createdAt: { type: string }

    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice]
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }

    OrderInput:
      type: object
      required: [customerName, customerPhone, customerCity, customerAddress, items]
      properties:
        customerName: { type: string, minLength: 1 }
        customerPhone: { type: string, minLength: 1 }
        customerCity: { type: string, minLength: 1 }
        customerAddress: { type: string, minLength: 1 }
        notes: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"

    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer, minimum: 1 }

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }

    OrderList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    CsvExport:
      type: object
      required: [csv]
      properties:
        csv: { type: string }

    # ── Admin Auth ────────────────────────────────────────────────────────────
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }

    AdminSession:
      type: object
      required: [username, authenticated]
      properties:
        username: { type: string }
        authenticated: { type: boolean }

    # ── Stats ─────────────────────────────────────────────────────────────────
    StatsSummary:
      type: object
      required: [totalOrders, totalRevenue, totalProducts, pendingOrders, recentOrders]
      properties:
        totalOrders: { type: integer }
        totalRevenue: { type: number }
        totalProducts: { type: integer }
        pendingOrders: { type: integer }
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"

    TopProduct:
      type: object
      required: [productId, productName, totalOrdered]
      properties:
        productId: { type: integer }
        productName: { type: string }
        totalOrdered: { type: integer }

    OrderStatusCount:
      type: object
      required: [status, count]
      properties:
        status: { type: string }
        count: { type: integer }