# lib/api-zod/src/generated/types/productList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pottery & Artisan eCommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductList {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}